rename from bid history to transaction history to be more coherent
